Populating DVRStream methods
+ filling up methods in class DVRStream
+ Added DVRInfo
+ Updated main Application class

+++ dvr/src/org/red5/server/api/stream/data/DVRStreamInfo.java
@@ -10,7 +10,7 @@ public class DVRStreamInfo {
 	
 	public Object stopRec;
 	
-	public long maxLen;
+	public double maxLen;
 	
 	public long begOffset;
